backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.22.1
alembic==1.14.1
greenlet==3.1.1

# Redis
redis[hiredis]==5.2.1

# AI / LLM
openai==1.59.9
numpy==2.2.2
razorpay==1.4.2
langgraph==0.2.62
langchain-openai==0.3.2
langchain-core==0.3.34

# Observability
langfuse==2.58.2

# HTTP Client
httpx==0.28.1

# Utilities
pydantic==2.10.5
uuid6==2024.7.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
aiosmtplib==3.0.2

# Google Gemini (Prescription OCR)
google-generativeai>=0.8.0

# Twilio (Refill Calls)
twilio>=9.0.0
